requirements.txt
google-generativeai
python-dotenv
rich
langchain-google-genai
langchain-community
chromadb
